give the scanner default buffer allocation / destruction methods, use them in sim inits
