+++ PYME/Acquire/Hardware/pointscan_shim/pointscan_camera.py
@@ -83,19 +83,44 @@ class BaseScanner(object):
         """
         raise NotImplementedError
     
-    def allocate_buffers(self, n_buffers):
-        """queue up a number of single-frame buffers
-        Note that each channel will be slotted into the queue as a separate frame (for now)
-        Parameters
-        ----------
-        n_buffers : int
-            number of single-frame (XY) buffers to allocate
+    # def allocate_buffers(self, n_buffers):
+    #     """queue up a number of single-frame buffers
+    #     Note that each channel will be slotted into the queue as a separate frame (for now)
+    #     Parameters
+    #     ----------
+    #     n_buffers : int
+    #         number of single-frame (XY) buffers to allocate
         
-        """
-        self.free_buffers = queue.Queue()
-        self.full_buffers = queue.Queue()
-        self.n_full = 0
-        raise NotImplementedError
+    #     """
+    #     self.free_buffers = queue.Queue()
+    #     self.full_buffers = queue.Queue()
+    #     self.n_full = 0
+    #     raise NotImplementedError
+    
+    def allocate_buffers(self, n_buffers):
+            """queue up a number of single-frame buffers
+            Note that each channel will be slotted into the queue as a separate frame (for now)
+            Parameters
+            ----------
+            n_buffers : int
+                number of single-frame (XY) buffers to allocate
+            
+            """
+            self.free_buffers = queue.Queue()
+            self.full_buffers = queue.Queue()
+            self.n_full = 0
+            for ind in range(n_buffers):
+                self.free_buffers.put(np.zeros((self.width, self.height), 
+                                            dtype=self.dtype))
+        
+    def destroy_buffers(self):
+        with self.full_buffer_lock:
+            self.n_full = 0
+            while not self.full_buffers.empty():
+                try:
+                    self.full_buffers.get_nowait()
+                except queue.Empty:
+                    pass
     
     def stop(self):
         raise NotImplementedError
